new MapActivity and Dashboard is no longer an activity
The dashboard will now wrap our current activity content in order to get access to the dashboard at any time in the app.
Tests and other activities have been slightly changed accordingly. MapActivity is now the main activity launched on sign up
DashboardActivityTest has been commented out for now and will be modified in upcoming commits
