+++ app/src/main/AndroidManifest.xml
@@ -64,11 +64,6 @@
             android:label="@string/title_activity_edit_profile"
             android:theme="@style/Theme.CoachMe" />
         <activity
-            android:name=".DashboardActivity"
-            android:exported="false"
-            android:label="@string/title_activity_dashboard"
-            android:theme="@style/Theme.CoachMe" />
-        <activity
             android:name=".SelectSportsActivity"
             android:exported="true"
             android:label="@string/title_activity_multiselect_list"
@@ -78,6 +73,11 @@
             android:exported="true"
             android:label="@string/title_activity_chat"
             android:windowSoftInputMode="adjustResize"
+            android:theme="@style/Theme.CoachMe" />
+        <activity
+            android:name=".map.MapActivity"
+            android:exported="false"
+            android:label="@string/title_activity_map"
             android:theme="@style/Theme.CoachMe" />
     </application>
